+++ src/app/screens/Trace/styles.module.scss
@@ -1,0 +1,99 @@
+@import 'scss/variables/_mediaQueries.scss';
+
+.link {
+  position: absolute;
+  top: 2%;
+  left: 2%;
+  font-size: 20px;
+  color: black;
+}
+
+.link:hover {
+  font-size: 22px;
+  color: blue;
+  text-decoration: none;
+}
+
+.container {
+  max-height: 100%;
+  align-items: center;
+  @media #{$mobile} {
+    display: flex;
+    flex-direction: column-reverse;
+  }
+}
+
+.containerTrace {
+  @media #{$desktop}, #{$tablet} {
+    background-color: rgba(200, 200, 200, 1);
+    border-radius: 25px;
+    color: white;
+    width: 25%;
+    float: left;
+    display: flex;
+    justify-content: center;
+  }
+  @media #{$mobile} {
+    background-color: rgba(200, 200, 200, 1);
+    border-radius: 25px;
+    color: white;
+    width: 80%;
+    display: flex;
+    justify-content: center;
+  }
+}
+
+.trace {
+  background-color: white;
+}
+
+.containerCalculator {
+  @media #{$desktop} {
+    width: 75%;
+    border-radius: 50px;
+    background-color: rgba(241, 135, 48, 0.8);
+    float: right;
+  }
+  @media #{$tablet-landscape} {
+    width: 75%;
+    border-radius: 50px;
+    background-color: rgba(241, 135, 48, 0.8);
+    float: right;
+  }
+  @media #{$tablet-portrait} {
+    width: 75%;
+    border-radius: 50px;
+    background-color: rgba(241, 135, 48, 0.8);
+    float: right;
+  }
+  @media #{$mobile} {
+    width: 100%;
+    height: 100%;
+    background-color: rgba(241, 135, 48, 0.8);
+    border-radius: 10px;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    border: 1px solid black;
+  }
+}
+
+.input {
+  height: 20%;
+  background-color: rgba(110, 110, 110, 1);
+  padding: 3% 15%;
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  @media #{$mobile} {
+    padding: 2% 15%;
+  }
+}
+
+.buttons {
+  @media #{$desktop, $tablet} {
+    padding: 5% 24%;
+  }
+  @media #{$mobile} {
+    padding: 0;
+  }
+}
