+++ src/constants/routes.js
@@ -1,2 +1,2 @@
-// export const HOME = '/';
-// export const QUOTES = '/quotes';
+export const HOME = '/';
+export const TRACE = '/trace';
